+++ support/multi_input/json/codeGo/path1DataReader.go
@@ -18,7 +18,7 @@ func (this *Path1DataReader) FileName() FileName {
 
 func (this *Path1DataReader) FromData(data []byte) error {
 	this.Path1DataStorer = &Path1DataStorer{
-		Datas: map[int64]*Path1Data{},
+		Datas: map[int32]*Path1Data{},
 	}
 
 	if err := json.Unmarshal(data, this.Path1DataStorer); err != nil {
@@ -30,7 +30,7 @@ func (this *Path1DataReader) FromData(data []byte) error {
 
 func (this *Path1DataReader) MergeData(data []byte) error {
 	tmpl := &Path1DataStorer{
-		Datas: map[int64]*Path1Data{},
+		Datas: map[int32]*Path1Data{},
 	}
 
 	if err := json.Unmarshal(data, tmpl); err != nil {
@@ -39,7 +39,7 @@ func (this *Path1DataReader) MergeData(data []byte) error {
 
 	if this.Path1DataStorer == nil {
 		this.Path1DataStorer = &Path1DataStorer{
-			Datas: map[int64]*Path1Data{},
+			Datas: map[int32]*Path1Data{},
 		}
 	}
 
@@ -58,12 +58,12 @@ func (this *Path1DataReader) Clear() {
 	this.Path1DataStorer = nil
 }
 
-func (this *Path1DataReader) Get(key int64) (result *Path1Data, ok bool) {
+func (this *Path1DataReader) Get(key int32) (result *Path1Data, ok bool) {
 	result, ok = this.Datas[key]
 	return result, ok
 }
 
-func (this *Path1DataReader) Keys() (result []int64) {
+func (this *Path1DataReader) Keys() (result []int32) {
 	for itor := range this.Datas {
 		result = append(result, itor)
 	}
