+++ support/multi_input/proto/codeGo/path3DataReader.go
@@ -19,7 +19,7 @@ func (this *Path3DataReader) FileName() FileName {
 
 func (this *Path3DataReader) FromData(data []byte) error {
 	this.Path3DataStorer = &Path3DataStorer{
-		Datas: map[int64]*Path3Data{},
+		Datas: map[int32]*Path3Data{},
 	}
 
 	if err := proto.Unmarshal(data, this.Path3DataStorer); err != nil {
@@ -31,7 +31,7 @@ func (this *Path3DataReader) FromData(data []byte) error {
 
 func (this *Path3DataReader) MergeData(data []byte) error {
 	tmpl := &Path3DataStorer{
-		Datas: map[int64]*Path3Data{},
+		Datas: map[int32]*Path3Data{},
 	}
 
 	if err := proto.Unmarshal(data, tmpl); err != nil {
@@ -40,7 +40,7 @@ func (this *Path3DataReader) MergeData(data []byte) error {
 
 	if this.Path3DataStorer == nil {
 		this.Path3DataStorer = &Path3DataStorer{
-			Datas: map[int64]*Path3Data{},
+			Datas: map[int32]*Path3Data{},
 		}
 	}
 
@@ -59,12 +59,12 @@ func (this *Path3DataReader) Clear() {
 	this.Path3DataStorer = nil
 }
 
-func (this *Path3DataReader) Get(key int64) (result *Path3Data, ok bool) {
+func (this *Path3DataReader) Get(key int32) (result *Path3Data, ok bool) {
 	result, ok = this.Datas[key]
 	return result, ok
 }
 
-func (this *Path3DataReader) Keys() (result []int64) {
+func (this *Path3DataReader) Keys() (result []int32) {
 	for itor := range this.Datas {
 		result = append(result, itor)
 	}
